% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessibility.R
\name{killer_luminance}
\alias{killer_luminance}
\title{WCAG relative luminance}
\usage{
killer_luminance(colours)
}
\arguments{
\item{colours}{A character vector of colours.}
}
\value{
A numeric vector of relative luminances, between 0 and 1.
}
\description{
WCAG relative luminance
}
\examples{
killer_luminance(c("white", "black", "#FF524F"))
}
