% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{killer_names}
\alias{killer_names}
\title{Names of all available palettes}
\usage{
killer_names()
}
\value{
A character vector of palette names, in album release order, with
\code{"greatest_hits"} last.
}
\description{
Names of all available palettes
}
\examples{
killer_names()
}
\seealso{
\code{\link[=killer_palette_info]{killer_palette_info()}} for names alongside album and year.
}
