% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pal.R
\name{plot.killer_palette}
\alias{plot.killer_palette}
\title{Plot a palette as swatches}
\usage{
\method{plot}{killer_palette}(x, ...)
}
\arguments{
\item{x}{A \code{killer_palette}, as returned by \code{\link[=killer_pal]{killer_pal()}}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly. Called for its side effect of drawing a plot.
}
\description{
Plot a palette as swatches
}
\examples{
plot(killer_pal("all_that_jazz"))
plot(killer_pal("miracle", n = 9, type = "diverging"))
}
